[perl #123817] Assert fail with attr in anon hash
These two one-liners were failing an assertion:

$y = { my $x : m }
$y = [ (my $x : m) ]

The lexical with an attribute was turning into an op of type ‘list’
with the OPpLVAL_INTRO flag set.  Many op constructors convert an
operand that is of type ‘list’ into the correct type, instead of
creating a new op to wrap around it.  Other ops were not expecting the
flag to be set, causing an assertion failure.  The easiest solution is
to turn off the flag in op_convert_list, which is where the op type
changes.  (I tried modifying other op types to accept the flag, but
that caused problems for B::Deparse, which inquires of B::Op_private
which ops can legitimately take that flag.)