[perl #123821] Fix assert fail with \(&$0)=0
If the operand of a reference constructor that is assigned to is not
valid, then S_lvref (in op.c) queues an error and does not bother to
convert the op into an lvref op.  This causes problems, since we have
already set lvref-specific flags on the op in the expectation that its
type will change.  We get assertion failures later when freeing the
op, because of the invalid flags.

Instead of bailing out when we have an error like this, just continue
processing it, as we do in many other places.