[perl #123848] Allow OPpTARGET_MY on qr
// in the scope of lexical $_ gets the OPpTARGET_MY flag set.  If it
is used as an operand to smartmatch, it gets converted to a qr op.
There is no need to turn off the flag, since it is harmless at run
time.  But we need to allow the flag on this op type to avoid asser-
tion failures when it is freed.