remove obsolete VOL in S_parse_body
commit 312caa8e97 removed dJMPENV from perl_parse_body (former perl_parse)
when it created perl_parse_body and perl_parse kept the dJMPENV, but this
commit forgot to remove the VOL tag on var dosearch. Since the VOL is
only required when the function has a setjmp in it, and this function
doesn't have a setjmp in it anymore, remove the VOL tag.