skip t/op/dump.t under Address Sanitizer
On my Fedora 20 system, when a process built with "clang
-fsanitize=address" coredumps, some clever root process called
abrt-install-ccpp-hook spends hours reading in zillions of zero bytes from
stdin. I guess this is due to ASan storing instrumentation data at very high
addresses, so the memory image of the dumping process consists of
terrabytes of intervening zeroes.

Whether the abrt system is doing doing something useful or is merely
buggy, the net effect is that t/op/dump.t (which purposely does a
coredump) causes the entire test harness to hang.

This commit just makes t/op/dump.c skip tests if -fsanitize is in
$Config{ccflags}.