Fix/simplify VMS-specific info in perlport.
It turns out that accumulating hastily-written patches for 20+
years does not make for the greatest reading experience, not to
mention the promises still being made that hadn't been true for
a long time, and the verbose explanations of things people really
don't need to worry about when running recent versions of Perl
on VMS versions released in the last 10 or 15 years.

Thanks to kwh for pointing out one of the more incoherent sentences,
which prompted me to take a closer look.