Fix regression in 5.21: /[A-Z]/ai
/[A-Z]/ai should match KELVIN SIGN, as it folds to a 'k'.  It should not
match under /aai, as that restricts fold matching.  But I tested for the
wrong symbol which ended up forbidding both /ai and /aai.

This commit changes to the correct symbol.  I also reordered the 'if'
while I was at it as a nano optimisation, to test for the /aa last, as
that is the less common part of the '&&' test.