cope with versions that do not match x.yyyzzz in corelist
Without this, we might be asked about 5.020 but be unable to climb
the delta tree because there is no delta entry for 5.020, only for
5.020.

This is a poor solution for now, because (for example) some versions
are not stored in x.yyyzzz format.  One is eveni n x.yyyzz!  We
should store things in a normalized format, and normalize arguments
on input.

For now, this will do.

(cherry picked from commit 4e70aa16a2f45d85bf4bef335d98c6a13e544bcb)