do not index the modules under lib/perl5db
These are there for testing, and there is nothing to index
under lib/perl5db anyway.  This has, in the past, led to junk
in 02packages.

I think this is probably not the perfect way to do this, but
it works and also does not seem like the worst way...