PATCH: [perl #124348] re/pat_advanced solaris failure
Tony Cook traced this down to a compiler bug.  But it's easy to change
the code to avoid the problem.  The expression evaluates to 0; and was
only in the form that caused the failure to document what was going on.
Now, instead the failed form is shown in comments, and 0 is used
directly.