perlunicode: Update nonchars discussion for Unicode 7.0
Unicode 7.0 changed the prohibition of noncharacters to merely "not
recommend" their use.  Perl continues to forbid them in strict input
checking (otherwise security issues could arise), but the discussion
about them needs to be updated to correspond with their new status.

The message raised when they are used probaby should change
correspondingly, but it is too late for 5.22 for that.

This commit deletes some text elsewhere about the noncharacter code
points.  This text really wasn't germane to a discussion about UTF-8
(wherein it appeared), as the encoding is irrelevant to these code
points.  They're not recommended in any UTF format.

Unicode spells the term "noncharacter" without a hyphen.  This pod
changes to follow that spelling.