perldelta: Highlight some 5.21 areas; fix others
I added XXX to lines where the text for this perldelta look like it
ought to change somewhat due to this being the delta from 5.20.2 to
5.22.0, instead of from one 5.21 release to another.  I believe, but am
uncertain that several of these should just be deleted; while others
require rewording.  Rather than not call out attention to them, I am
marking them to draw attention to them for people more expert in these
areas to adddress.

Several areas seemed clear to me that a bug introduced in some 5.21
version was fixed in a later 5.21 version, so has zero relevance for
making the jump from 5.20 to 5.22.  I just deleted those.

And I had minor rewording in others, where it mentioned 5.21.  We've
decided not to talk about development releases in the pods.