lib/utf8.t: Add some tests to stress EBCDIC
The tests for the Latin1 \xFF aren't a fair test of UTF-8 on EBCDIC
platforms, because it is generally a UTF-8 invariant character, so is
the same regardless of being in UTF-8 or not.  This adds some tests
where the UTF-EBCDIC version is 2 bytes (as well as the UTF-8 version).