Allow Perl to compile and work on Unicode releases without U+1E9E
U+1E9E LATIN CAPITAL LETTER SHARP S is handled specially by Perl,
because of its relationship to the infamous LATIN SMALL LETTER SHARP S,
which folds to 'ss', being the only character whose code point is < 256
to have a multi char fold, and this creates lots of special cases.

But U+1E9E wasn't in all Unicode releases.  Because Perl is supposed to
work with any release, we need to be able to compile when this character
isn't defined.  In some of those cases we use U+017F (LATIN SMALL LETTER
LONG S instead, which is in all releases.