perldelta: Rmv reference to internal flag
SCF_DO_SUBSTR is a flag internal to the current implementation of the
regular expression optimizer.  There is no need to proclaim its
existence to the outside world, and is just extraneous noise.

I myself do not understand this flag, and I've spent more time looking
at this code than all but a few people likely to be reading this
perldelta.  If someone who does understand it could explain it to me, I
would add comments to the code (after the freeze) to aid future readers.