Handle variant build product extensions in configure.com.
This addresses [perl #125284].  Object files, object libraries,
shared libraries, and executables can be given architecture-specific
extensions on VMS to allow for an installation directory that can
be shared by multiple platforms.  I guess this feature seemed like
a good idea twenty years ago when disks were much smaller, but it
hasn't been working for some time because, as noted in the ticket
some configure-time compilation was inconsistent in its use of the
alternate extensions.  There were a couple more in addition to
those mentioned in the ticket.

N.B.  t/TEST still won't run as it's looking explicitly for
perl.exe so that's a TODO.