mktables: Fix logic error
When I wrote this some years ago, I committed a logic error, so that a
constraint I thought applied was in fact somewhat different than what I
thought.  The new comments give the proper constraint.

The code is taking the data furnished by unicode.org and massaging it
into tables usable by the perl interpreter.  This bug can cause wrong
tables to be generated.  Fortunately, this bug does not arise with
typical unicode.org data sets, so it hasn't shown up as a problem.
Fixing it makes no difference in the current Unicode version.  I
discovered it only in compiling Unicode 1.1.5, somewhat tweaked by me,
so not the official data there either.